created very basic test case for the code
It is failing the test case (returning a vector of NaN's), but it is best to implement tree-to-latex parsing before debugging because reading the tree structure is unreasonably painful. Thus awaiting implementation of tree-to-latex parsing before continuing with the debugging of this code. The issue *seems* to be with the building of the Hessian matrix, specifically when differentiating something from the partial_derivatives vector. However, this is only my hunch